Fixed a problem with byte switching
This fixes an issue with byte switching. The preprocessor directive
used to determine the endianess of the system only worked on Linux
resulting in voice running over TCP only when umurmur runs on other
platforms. Note, this patch could and should be implemented better
but is working and should be regarded as a hotfix.