added support for building umurmur-monitor
It is now possible (with cmake) to build umurmur-monitor as a
subproject of umumur. To do so, use the following command line to clone
umurmur:

  git clone --recursive <url-to-repo>

or simply update your current copy (fetch, rebase, merge, whatever you
prefer) and then run:

  git submodule udate --init

Afterwards, create your build directory and use either ccmake or a
command line similiar to the following:

  cmake <path-to-CMakeLists.txt-of-umurmur> -DUSE_SHAREDMEMORY_API=ON \
        -DBUILD_UMURMUR_MONITOR=ON <other-flags>

Currently, building umurmur-monitor is OFF by default, even if cmake
finds the sources for the project. This might change in the future when
demanded.