Add show-addresses option, to optionally hide user's IP addresses
Currently IP addresses of every connected user are visible to everybody
through the user information dialog. This adds an option of whether or
not to show them, if set to true (default) then everything will be as
before, whereas with false they will not be shown to anybody.

IP addresses will still be logged in log messages and so forth.

Fixes #81.