Add -s flag to force switching users
As @C4K3 suggested in issue #105, the -s flag can be used to force
switching to the configured user and group combination without
daemonizing. Since it is best practice to not daemonize when running as
a systemd service, this flag, in combination with -d, allows systemd
users to switch users without daemonizing. Something like the following
would then be used in ExecStart:

  /usr/bin/umurmur -c /etc/umurmur/umurmur.conf -d -s