Fix compiling with autotools and openssl 1.1.0
Since the functions previously checked against have been removed in
openssl 1.1.0, it would not compile. This changes it so that functions
available in both 1.0.2 and 1.1.0 are used instead.

I originally had a lot of problems getting it to work, using the example
from https://wiki.openssl.org/index.php/Library_Initialization#Autoconf
as then it would no longer link against ssl. It turns out that this was
because when AC_CHECK_LIB has a action-if-found specified, it will NOT
add ssl to LIBS, as opposed to what is said on
https://autotools.io/autoconf/finding.html -- I discovered this after
finally just going to the original documentation, which correctly
documents the macro
(https://www.gnu.org/software/autoconf/manual/autoconf-2.61/html_node/Libraries.html)