Install config to CMAKE_INSTALL_SYSCONFDIR
CMakeLists.txt:
Include GNUInstallDirs and use CMAKE_INSTALL_SYSCONFDIR to install the
example configuration (defaults to /etc/umurmur/umurmur.conf). The
configuration is installed with 0640 permissions as it may contain
passwords and therefore should not be world readable.
Remove prior checks to the system for umurmur.conf as the merging or
backup of configuration files is historically taken care of by a package
management system on a target system.