Namespace default configuration file location
src/config.h.in:
Namespace the default configuration file location with the project's
name "umurmur", as in that location also the required certificates are
expected by default and using a namespaced location makes sense as to
not clutter the SYSCONFDIR (e.g. /etc/) with more files, as a namespace
is necessary anyways.

CMakeLists.txt:
Install the default configuration file to a "umurmur" namespaced
location.

Fixes #169